Issue 81 drop unique constraints postgres

+++ pkg/database/postgres/index.go
@@ -10,6 +10,9 @@ import (
 )
 
 func RemoveIndexStatement(tableName string, index *types.Index) string {
+	if index.IsUnique {
+		return fmt.Sprintf("alter table %s drop constraint %s", tableName, pq.QuoteIdentifier(index.Name))
+	}
 	return fmt.Sprintf("drop index %s", pq.QuoteIdentifier(index.Name))
 }
